Done api rating and random reader

+++ src/main/java/org/example/be_eproject_sem4/Dto/Auth/UserDto.java
@@ -24,7 +24,7 @@ public class UserDto {
     private String role;
     private Date birthDate;
     private boolean isVerified;
-    private int eloScore;
+    private double eloScore;
     private String bio;
     private String profilePicture;
     private String qrCode;
